+++ CHANGES.rst
@@ -13,6 +13,13 @@ divided into four categories.
 Read `r/changelog <http://www.reddit.com/r/changelog>`_ to be notified of
 upstream changes.
 
+Unreleased
+----------
+ * **[FEATURE]** Add arguments syntax and time to
+   :meth:`~praw.__init__.UnauthenticatedReddit.search`.
+ * **[BUGFIX]** Fix bug that prevented the usage of
+   :meth:`~praw.objects.Subreddit.search` when called from :obj:`.Subreddit`.
+
 PRAW 2.1.4
 ----------
  * **[FEATURE]** :meth:`~praw.__init__.ModOnlyMixin.get_mod_mail` can now be